Add changelog script
This script can be used to help us when generating change logs in Github. It will examine all merge requests that were made into master between two branches or tags within the current working directory. It will then perform a request to the Jira API to get additional information about the changes from an associated ticket, assuming the branch name has the associated ticket id in it (ex: DLTP-123-cool-feature). Note, this has been made better over time while I've been using it, but it's not the greatest at handling errors or validating user input. I'm hoping we can rewrite this in the future to be a bit more robust, and probably not be in bash.