Make VERSION optional
Adding a VERSION file doesn't always make sense, and in some cases the master branch is protected from direct pushes. This change makes it optional to create the version file to better support other projects that follow the same versioning schema, but don't require a file to track it. Additionally, this changes it to be more explicit about what tag should be pushed, just in case the developer has extras on their local git.