Bugfix when remote tags don't match local
If the remote has newer tags than local, the current implementation will try to recreate the tag, then fails to push. This changes to list the tags on the remote instead of local in order to calculate the new identifier. I chose this instead of 'git fetch --tags', just to avoid any weirdness with pulling tags for someone if they don't want to.